DEMOS-263 rename HeaderLower to default intent, lint
