+++ client/src/mock-data/userMocks.ts
@@ -1,6 +1,6 @@
 import { User } from "demos-server";
 import { GET_ALL_USERS, GET_USER_BY_ID } from "hooks/useUserOperations";
-import { HEADER_LOWER_QUERY } from "components/header/HeaderLower";
+import { HEADER_LOWER_QUERY } from "components/header/DefaultHeaderLower";
 import { PROFILE_BLOCK_QUERY } from "components/header/ProfileBlock";
 import { MockedResponse } from "@apollo/client/testing";
